+++ target-alpha/translate.c
@@ -2509,7 +2509,7 @@ CPUAlphaState * cpu_alpha_init (const char *cpu_model)
 }
 
 void gen_pc_load(CPUState *env, TranslationBlock *tb,
-                unsigned long searched_pc, int pc_pos, void *puc)
+                uintptr_t searched_pc, int pc_pos, void *puc)
 {
     env->pc = gen_opc_pc[pc_pos];
 }
